+++ R/survival_eval.R
@@ -220,9 +220,32 @@ calc_surv_from_prob <- function(x) {
 #' @inherit compute_surv
 #' @name eval_surv
 #' @keywords internal
-eval_surv <- function(x, time, cycle_length = 1,
-                       type = c("prob","surv"), ...) {
+eval_surv <- function(x, time, ...) {
   UseMethod("eval_surv")
+}
+
+#' @inherit compute_surv
+#' @name eval_surv
+#' @keywords internal
+compute_surv_ <- function(x, time, 
+                          cycle_length = 1, 
+                          type = c("prob", "survival"), ...){
+  type <- match.arg(type)
+  
+  if (type == "prob") {
+    time_ = c(time[1] - 1, time)
+  } else {
+    time_ = time
+  }
+  
+  check_cycle_inputs(time_, cycle_length)
+  
+  ret <- eval_surv(x, cycle_length * time_, ...)
+  if (type == "prob") {
+    # Calculate per-cycle failure prob
+    ret <- calc_prob_from_surv(ret)
+  }
+  ret
 }
 
 #' Evaluate Survival Distributions
@@ -247,30 +270,16 @@ eval_surv <- function(x, time, cycle_length = 1,
 #'   conditional probabilities of event for each cycle.
 #' @export
 compute_surv <- memoise::memoise(
-  eval_surv,
+  compute_surv_,
   ~ memoise::timeout(options()$heemod.memotime)
 )
 
 #' @rdname eval_surv
 #' @export
-eval_surv.survfit <- function(x, time, cycle_length = 1,
-                               type = c("prob","surv"), ...) {
+eval_surv.survfit <- function(x, time,  ...) {
   
   dots <- list(...)
   
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
-  
-  time_surv <- cycle_length * time_
-  
-  # Extract the product-limit tables for all strata
   pl_table <- extract_strata(x)
   
   # Identify the terms which separate groups (if any)
@@ -288,11 +297,11 @@ eval_surv.survfit <- function(x, time, cycle_length = 1,
       maxtime <- max(d$time)
       selector <- (time > maxtime)
       # Use stepfun to look up survival probabilities
-      value <- stats::stepfun(d$time[-1], d$surv)(time_surv)
-      # Use NA when time_surv > max time_surv
+      value <- stats::stepfun(d$time[-1], d$surv)(time)
+      # Use NA when time > max time
       value[selector] <- as.numeric(NA)
       tibble::data_frame(
-        t = time_surv,
+        t = time, 
         value = value,
         n = d$n[1])
     }
@@ -304,6 +313,7 @@ eval_surv.survfit <- function(x, time, cycle_length = 1,
     }
     # If covariates are not provided, do weighted average for each time.
     agg_df <- surv_df %>%
+      tibble::as_tibble() %>% 
       dplyr::group_by_(~ t) %>%
       dplyr::summarize_(value = ~ sum(value * n) / sum(n))
   } else {
@@ -319,34 +329,16 @@ eval_surv.survfit <- function(x, time, cycle_length = 1,
   
   # Get the vector of predictions
   ret <- agg_df$value
-  
-  if (type == "prob") {
-    # Calculate per-cycle failure prob
-    ret <- calc_prob_from_surv(ret)
-  }
-  
   return(ret)
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.flexsurvreg <- function(x, time, cycle_length = 1,
-                                   type = c("prob", "surv"), ...) {
+eval_surv.flexsurvreg <- function(x, time,  ...) {
   
   dots <- list(...)
   
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
-  
-  time_surv <- cycle_length * time_
-  
+  time_surv <- time
   # Extract parameter estimates
   coef_obj <- x$coefficients
   
@@ -408,24 +400,16 @@ eval_surv.flexsurvreg <- function(x, time, cycle_length = 1,
   # Just get the results column
   ret <- surv_df$value
   
-  if (type == "prob") {
-    # Calculate per-cycle failure prob
-    ret <- calc_prob_from_surv(ret)
-  }
-  
   return(ret)
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_model <- function(x, time, cycle_length = 1,
-                                  type = c("prob", "surv"), ...) {
+eval_surv.surv_model <- function(x, time,  ...) {
   eval_surv(
     x$dist,
     time = time,
-    cycle_length = cycle_length,
     covar = x$covar,
-    type = type,
     ...
   )
   
@@ -433,80 +417,43 @@ eval_surv.surv_model <- function(x, time, cycle_length = 1,
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_projection <- function(x, time,
-                                       cycle_length = 1,
-                                       type = c("prob", "surv"), ...) {
-  
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
-  
-  time_surv <- cycle_length * time_
-  
-  ret <- numeric(length(time_))
+eval_surv.surv_projection <- function(x, time, ...) {
+
+  ret <- numeric(length(time))
   
   surv1 <- eval_surv(
     x$dist1,
-    time = time_,
-    cycle_length = cycle_length,
-    type = "surv"
+    time = time 
   )
   surv2 <- eval_surv(
     x$dist2,
-    time = time_,
-    cycle_length = cycle_length,
-    type = "surv"
+    time = time
   )
   
-  ind_s1 <- time_surv < x$at
-  ind_s2 <- time_surv >= x$at
+  ind_s1 <- time < x$at
+  ind_s2 <- time >= x$at
   
   surv1_p_at <- eval_surv(
     x$dist1,
-    time = 1,
-    cycle_length = x$at,
-    type = "surv")
+    time = x$at 
+    )
   surv2_p_at <- eval_surv(
     x$dist2,
-    time = 1,
-    cycle_length = x$at, 
-    type = "surv",
+    time = x$at,
     .internal = TRUE)
   
   ret[ind_s1] <- surv1[ind_s1]
   ret[ind_s2] <- (surv2 * surv1_p_at / surv2_p_at)[ind_s2]
-  
-  if(type == "prob") {
-    ret <- calc_prob_from_surv(ret)
-  }
   
   ret
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_pooled <- function(x, time,
-                                   cycle_length = 1,
-                                   type = c("prob", "surv"), ...) {
-  
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
+eval_surv.surv_pooled <- function(x, time, ...) {
   
   # Determine dimensions of matrix and initialize
-  n_cycle <- length(time_)
+  n_cycle <- length(time)
   n_dist <- length(x$dists)
   surv_mat <- matrix(nrow = n_cycle, ncol = n_dist)
   
@@ -516,8 +463,7 @@ eval_surv.surv_pooled <- function(x, time,
     surv_mat[ ,i] <- x$weights[i] / sum(x$weights) *
       eval_surv(
         x$dists[[i]],
-        time = time_,
-        cycle_length = cycle_length,
+        time = time, 
         type = "surv"
       )
   }
@@ -525,125 +471,76 @@ eval_surv.surv_pooled <- function(x, time,
   # Calculate weighted average as the row sums
   ret <- rowSums(surv_mat)
   
-  if (type == "prob") {
-    ret <- calc_prob_from_surv(ret)
-  }
-  
   ret
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_ph <- function(x, time,
-                               cycle_length = 1,
-                               type = c("prob", "surv"), ...) {
-  
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
+eval_surv.surv_ph <- function(x, time, ...) {
   
   ret <- eval_surv(
     x$dist,
-    time = time_,
-    cycle_length = cycle_length,
-    type = "surv"
+    time = time 
   ) ^ x$hr
   
-  if (type == "prob") {
-    ret <- calc_prob_from_surv(ret)
-  }
-  
   ret
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_aft <- function(x, time,
-                                cycle_length = 1,
-                                type = c("prob", "surv"), ...) {
+eval_surv.surv_shift <- function(x, time, ...) {
   
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
+  time_ <- time
+  time_ <- time_ - x$shift
+  ret <- rep(1, length(time_))
+  keep_me <- time_ >= 0
+  if(any(keep_me)){
+    time_ <- time_[keep_me]
+    ##check_cycle_inputs(time_, cycle_length)
+    ret[keep_me] <- eval_surv(
+      x$dist,
+      time = time_ 
+    ) 
   }
-  
-  check_cycle_inputs(time_, cycle_length)
+
+  ret
+}
+
+
+#' @rdname eval_surv
+#' @export
+eval_surv.surv_aft <- function(x, time, ...) {
   
   ret <- eval_surv(
     x$dist,
-    time = time_,
-    cycle_length = cycle_length/x$af,
-    type = "surv"
+    time = time/x$af 
   )
-  
-  if (type == "prob") {
-    ret <- calc_prob_from_surv(ret)
-  }
   
   ret
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_po <- function(x, time,
-                               cycle_length = 1,
-                               type = c("prob", "surv"), ...) {
+eval_surv.surv_po <- function(x, time, ...) {
   
   dots <- list(...)
-  
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
-  
+
   p <- eval_surv(
     x$dist,
-    time = time_,
-    cycle_length = cycle_length,
-    type = "surv"
+    time = time
   )
   
   ret <- 1 / ((((1 - p) / p) * x$or) + 1)
   
-  if (type == "prob") {
-    ret <- calc_prob_from_surv(ret)
-  }
-  
   ret
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_add_haz <- function(x, time,
-                                    cycle_length = 1,
-                                    type = c("prob", "surv"), ...) {
-  
-  type <- match.arg(type)
-  
-  if (type == "prob") {
-    time_ = c(time[1] - 1, time)
-  } else {
-    time_ = time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
+eval_surv.surv_add_haz <- function(x, time, ...) {
   
   # Determine dimensions of matrix and initialize
-  n_cycle <- length(time_)
+  n_cycle <- length(time)
   n_dist <- length(x$dists)
   surv_mat <- matrix(nrow = n_cycle, ncol = n_dist)
   
@@ -652,39 +549,20 @@ eval_surv.surv_add_haz <- function(x, time,
   for (i in seq_len(n_dist)) {
     surv_mat[ ,i] <- eval_surv(
       x$dists[[i]],
-      time = time_,
-      cycle_length = cycle_length,
-      type = "surv"
+      time = time 
     )
   }
   
   # Apply independent risks
   ret <- apply(surv_mat, 1, function(z) prod(z))
   
-  if (type == "prob") {
-    ret <- calc_prob_from_surv(ret)
-  }
-  
   ret
 }
 
 #' @rdname eval_surv
 #' @export
-eval_surv.surv_dist <- function(x, time,
-                                 cycle_length = 1,
-                                 type = c("prob", "surv"), ...) {
-  type <- match.arg(type)
-  
-  if(type == "prob") {
-    time_ <- c(time[1] - 1, time)
-  }else {
-    time_ <- time
-  }
-  
-  check_cycle_inputs(time_, cycle_length)
-  
-  times_surv <- cycle_length * time_
-  
+eval_surv.surv_dist <- function(x, time, ...) {
+
   if (! requireNamespace("flexsurv")) {
     stop("'flexsurv' package required.")
   }
@@ -693,13 +571,9 @@ eval_surv.surv_dist <- function(x, time,
             envir = asNamespace("flexsurv"))
   
   args <- x[- match("distribution", names(x))]
-  args[["q"]] <- times_surv
+  args[["q"]] <- time
   args[["lower.tail"]] <- FALSE
   ret <- do.call(pf, args)
-  
-  if (type == "prob") {
-    ret <- calc_prob_from_surv(ret)
-  }
   
   ret
 }